requirements.txt
aioice==0.9.0
aiortc==1.9.0
altair==5.5.0
attrs==24.2.0
av==12.3.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
dnspython==2.7.0
gitdb==4.0.11
GitPython==3.1.43
google-crc32c==1.6.0
idna==3.10
ifaddr==0.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
narwhals==1.14.2
numpy==1.26.4
opencv-python==4.10.0.84
packaging==24.2
pandas==2.2.3
pillow==10.4.0
protobuf==4.25.5
pyarrow==18.1.0
pycparser==2.22
pydeck==0.9.1
pyee==12.1.1
Pygments==2.18.0
pylibsrtp==0.10.0
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
pytz==2024.2
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.21.0
six==1.16.0
smmap==5.0.1
st-theme==1.2.3
streamlit==1.33.0
streamlit-navigation-bar==3.2.0
streamlit-webrtc==0.47.9
tenacity==8.5.0
toml==0.10.2
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
yt-dlp==2024.11.18
